+++ index.js
@@ -72,13 +72,13 @@ class merakiDevice {
     this.disableLogInfo = config.disableLogInfo;
     this.hideUnconfiguredSsids = config.hideUnconfiguredSsids;
     this.hideSsidByName = config.hideSsidByName || [];
-    this.getClientByName = config.getClientByName || [];
+    this.getClientByNameOrMac = config.getClientByNameOrMac || [];
 
     //get Device info
-    this.manufacturer = config.manufacturer || 'Cisco/Meraki';
-    this.modelName = config.modelName || 'Model Name';
-    this.serialNumber = config.serialNumber || 'Serial Number';
-    this.firmwareRevision = config.firmwareRevision || 'Firmware Revision';
+    this.manufacturer = 'Meraki';
+    this.modelName = this.name || 'Network Name';
+    this.serialNumber = this.networkId || 'Network ID';
+    this.firmwareRevision = this.organizationId || 'Organization ID';
 
     //setup variables
     this.checkDeviceInfo = false;
@@ -86,27 +86,27 @@ class merakiDevice {
     this.startPrepareAccessory = true;
 
     //data
-    this.merakiDashboardClientsData = {
+    this.dashboardClientsData = {
       'status': 0
     };
-    this.merakiMrSsidsData = {
+    this.ssidsData = {
       'status': 0
     };
 
     //meraki dashboard
-    this.merakiDashboardClientsCount = 0;
-    this.exposedClientByNameCount = this.getClientByName.length;
-    this.showClientsByName = false;
+    this.dashboardClientsCount = 0;
+    this.exposedClientByNameCount = this.getClientByNameOrMac.length;
 
     //meraki mr
-    this.allSsidsCount = 0;
+    this.ssidsCount = 0;
     this.exposedSsidsCount = 0;
     this.hiddenSsidsCount = this.hideSsidByName.length;
 
-    this.prefDir = path.join(api.user.storagePath(), 'meraki');
-    this.dashboardClientsUrl = this.host + '/api/v1/networks/' + this.networkId + '/clients?perPage=255';
+
+    //meraki url
+    this.dashboardClientsUrl = this.host + '/api/v1/networks/' + this.networkId + '/clients?perPage=255&timespan=1209600';
     this.mxUrl = this.host + '/api/v1/networks/' + this.networkId + '/appliance/ports';
-    this.mrSsidUrl = this.host + '/api/v1/networks/' + this.networkId + '/wireless/ssids';
+    this.ssidsUrl = this.host + '/api/v1/networks/' + this.networkId + '/wireless/ssids';
     this.msUrl = this.host + '/api/v1/devices/' + this.serialNumber + '/switch/ports';
     this.mvUrl = this.host + '/api/v1/devices/' + this.serialNumber + '/camera';
 
@@ -119,7 +119,10 @@ class merakiDevice {
       }
     });
 
-    //check if prefs directory ends with a /, if not then add it
+    //preferences directory
+    this.prefDir = path.join(api.user.storagePath(), 'meraki');
+
+    //check if prefs directory exist
     if (fs.existsSync(this.prefDir) == false) {
       fsPromises.mkdir(this.prefDir);
     }
@@ -127,44 +130,44 @@ class merakiDevice {
     //Check device state
     setInterval(function () {
       if (this.checkDeviceInfo || this.checkDeviceState) {
-        this.updateMerakiDashboardClientsData();
+        this.updateDashboardClientsData();
       }
     }.bind(this), this.refreshInterval * 1000);
 
-    this.updateMerakiDashboardClientsData();
+    this.updateDashboardClientsData();
   }
 
-  async updateMerakiDashboardClientsData() {
-    this.log.debug('Device: %s %s, requesting merakiDashboardClientsData.', this.host, this.name);
+  async updateDashboardClientsData() {
+    this.log.debug('Device: %s %s, requesting dashboardClientsData.', this.host, this.name);
     try {
-      const merakiDashboardClientsData = await this.meraki.get(this.dashboardClientsUrl);
-      this.log.debug('Debug merakiDashboardClientsData: %s', merakiDashboardClientsData.data[0]);
-      const merakiDashboardClientsCount = merakiDashboardClientsData.data.length;
+      const dashboardClientsData = await this.meraki.get(this.dashboardClientsUrl);
+      this.log.debug('Debug dashboardClientsData: %s', dashboardClientsData.data[0]);
+      const dashboardClientsCount = dashboardClientsData.data.length;
 
-      this.merakiDashboardClientsData = merakiDashboardClientsData;
-      this.merakiDashboardClientsCount = merakiDashboardClientsCount;
+      this.dashboardClientsData = dashboardClientsData;
+      this.dashboardClientsCount = dashboardClientsCount;
 
-      this.updateMerakiMrSsidsData();
+      this.updateSsidsData();
     } catch (error) {
-      this.log.error('Device: %s %s, merakiDashboardClientsData error: %s', this.host, this.name, error);
+      this.log.error('Device: %s %s, dashboardClientsData error: %s', this.host, this.name, error);
       this.checkDeviceState = false;
       this.checkDeviceInfo = true;
     };
   }
 
-  async updateMerakiMrSsidsData() {
-    this.log.debug('Device: %s %s, requesting merakiMrSsidsData.', this.host, this.name);
+  async updateSsidsData() {
+    this.log.debug('Device: %s %s, requesting ssidsData.', this.host, this.name);
     try {
-      const merakiMrSsidsData = await this.meraki.get(this.mrSsidUrl);
-      this.log.debug('Debug merakiMrData: %s', merakiMrSsidsData.data[0]);
-      const allSsidsCount = merakiMrSsidsData.data.length;
+      const ssidsData = await this.meraki.get(this.ssidsUrl);
+      this.log.debug('Debug merakiMrData: %s', ssidsData.data[0]);
+      const ssidsCount = ssidsData.data.length;
 
-      this.merakiMrSsidsData = merakiMrSsidsData;
-      this.allSsidsCount = allSsidsCount;
+      this.ssidsData = ssidsData;
+      this.ssidsCount = ssidsCount;
 
       const getDeviceInfo = !this.checkDeviceState ? this.getDeviceInfo() : this.updateDeviceState();
     } catch (error) {
-      this.log.error('Device: %s %s, merakiMrSsidsData error: %s', this.host, this.name, error);
+      this.log.error('Device: %s %s, ssidsData error: %s', this.host, this.name, error);
       this.checkDeviceState = false;
       this.checkDeviceInfo = true;
     };
@@ -193,96 +196,106 @@ class merakiDevice {
     this.log.debug('Device: %s %s, update device state.', this.host, this.name);
     try {
       //get devices data;
-      const merakiDashboardClientsData = this.merakiDashboardClientsData;
-      const merakiMrSsidsData = this.merakiMrSsidsData;
+      const dashboardClientsData = this.dashboardClientsData;
+      const ssidsData = this.ssidsData;
 
       //get devices data variables
-      const merakiDashboardClientsCount = this.merakiDashboardClientsCount;
+      const dashboardClientsCount = this.dashboardClientsCount;
       const exposedClientByNameCount = this.exposedClientByNameCount;
-      const allSsidsCount = this.allSsidsCount;
+      const ssidsCount = this.ssidsCount;
       const hiddenSsidsCount = this.hiddenSsidsCount;
 
 
-      //daschboards filter clients
-      if (merakiDashboardClientsData.status == 200) {
-        this.allClientsName = new Array();
+      //daschboards clients
+      if (dashboardClientsData.status == 200) {
         this.clientsId = new Array();
         this.clientsUser = new Array();
+        this.clientsMac = new Array();
+
         this.clientsPolicyMac = new Array();
         this.clientsPolicyPolicy = new Array();
         this.clientsPolicyGroupPolicyId = new Array();
         this.clientsPolicyState = new Array();
 
-        for (let i = 0; i < merakiDashboardClientsCount; i++) {
-          const clientDescription = merakiDashboardClientsData.data[i].description;
-          this.allClientsName.push(clientDescription);
+        const dashboardClientsDescription = new Array();
+        const dashboardClientsMac = new Array();
+        for (let i = 0; i < dashboardClientsCount; i++) {
+          const clientDescription = dashboardClientsData.data[i].description;
+          const clientMac = dashboardClientsData.data[i].mac;
+          dashboardClientsDescription.push(clientDescription);
+          dashboardClientsMac.push(clientMac);
 
           for (let j = 0; j < exposedClientByNameCount; j++) {
-            const showClientsByName = this.allClientsName.indexOf(this.getClientByName[j].name) >= 0);
+            const clientMode = this.getClientByNameOrMac[j].mode;
+            const clientNameOrMac = this.getClientByNameOrMac[j].name;
+            const clientsByNameOrMac = clientMode ? dashboardClientsMac : dashboardClientsDescription;
+            const getClient = (clientsByNameOrMac.indexOf(clientNameOrMac) >= 0);
 
-            if (showClientsByName) {
-              const clientId = merakiDashboardClientsData.data[i].id;
-              const clientUser = merakiDashboardClientsData.data[i].description;
+            if (getClient) {
+              const clientId = dashboardClientsData.data[i].id;
+              const clientUser = dashboardClientsData.data[i].description;
+              const clientMac = dashboardClientsData.data[i].mac;
 
               try {
-                const dashboardClientsByIdPolicyUrl = this.host + '/api/v1/networks/' + this.networkId + '/clients/' + clientId + '/policy';
-                const merakiDashboardClientPolicyData = await this.meraki.get(dashboardClientsByIdPolicyUrl);
-                this.log.debug('Debug merakiDashboardClientPolicyData: %s', merakiDashboardClientPolicyData.data);
+                const dashboardClientsPolicyUrl = this.host + '/api/v1/networks/' + this.networkId + '/clients/' + clientId + '/policy';
+                const dashboardClientsPolicyData = await this.meraki.get(dashboardClientsPolicyUrl);
+                this.log.debug('Debug dashboardClientsPolicyData: %s', dashboardClientsPolicyData.data);
 
-                const clientPolicyMac = merakiDashboardClientPolicyData.data.mac;
-                const clientPolicyPolicy = merakiDashboardClientPolicyData.data.devicePolicy;
-                const clientPolicyGroupPolicyId = merakiDashboardClientPolicyData.data.groupPolicyId;
+                const clientPolicyMac = dashboardClientsPolicyData.data.mac;
+                const clientPolicyPolicy = dashboardClientsPolicyData.data.devicePolicy;
+                const clientPolicyGroupPolicyId = dashboardClientsPolicyData.data.groupPolicyId;
                 const clientPolicyState = (clientPolicyPolicy == 'Normal' || clientPolicyPolicy == 'Whitelisted');
+
+                if (this.merakiClientPolicyServices) {
+                  this.merakiClientPolicyServices[j]
+                    .updateCharacteristic(Characteristic.On, clientPolicyState);
+                }
 
                 this.clientsPolicyMac.push(clientPolicyMac);
                 this.clientsPolicyPolicy.push(clientPolicyPolicy);
                 this.clientsPolicyGroupPolicyId.push(clientPolicyGroupPolicyId);
                 this.clientsPolicyState.push(clientPolicyState);
-                
-                if (this.merakiClientPolicyServices) {
-                    this.merakiClientPolicyServices[j]
-                    .updateCharacteristic(Characteristic.On, clientsPolicyState);
-                 }
               } catch (error) {
-                this.log.error('Device: %s %s, merakiDashboardClientPolicyData error: %s', this.host, this.name, error);
+                this.log.error('Device: %s %s, dashboardClientsPolicyData error: %s', this.host, this.name, error);
               }
               this.clientsId.push(clientId);
               this.clientsUser.push(clientUser);
+              this.clientsMac.push(clientMac);
             }
           }
         }
       }
 
 
-      //MR device
-      if (merakiMrSsidsData.status == 200) {
-        this.allSsidsName = new Array();
+      //SSIDs
+      if (ssidsData.status == 200) {
         this.ssidsName = new Array();
-        this.ssidsState = new Array();
+        this.exposedSsidsName = new Array();
+        this.exposedSsidsState = new Array();
+        this.hidenSsidsName = new Array();
 
-        this.hidenSsidsByName = new Array();
         for (let j = 0; j < hiddenSsidsCount; j++) {
           const hiddedSsidByName = this.hideSsidByName[j].name;
-          this.hidenSsidsByName.push(hiddedSsidByName);
+          this.hidenSsidsName.push(hiddedSsidByName);
         }
 
-        for (let i = 0; i < allSsidsCount; i++) {
-          const currentSsidName = merakiMrSsidsData.data[i].name;
-          this.allSsidsName.push(currentSsidName);
+        for (let i = 0; i < ssidsCount; i++) {
+          const ssidName = ssidsData.data[i].name;
+          this.ssidsName.push(ssidName);
 
-          const showConfiguredSsids = this.hideUnconfiguredSsids ? (currentSsidName.substr(0, 12) != 'Unconfigured') : true;
+          const showConfiguredSsids = this.hideUnconfiguredSsids ? (ssidName.substr(0, 12) != 'Unconfigured') : true;
           if (showConfiguredSsids) {
-            const ssidName = merakiMrSsidsData.data[i].name;
-            const ssidState = (merakiMrSsidsData.data[i].enabled == true)
+            const ssidName = ssidsData.data[i].name;
+            const ssidState = (ssidsData.data[i].enabled == true)
 
-            const pushName = (this.hidenSsidsByName.indexOf(ssidName) >= 0) ? false : this.ssidsName.push(ssidName);
-            const pushState = (this.hidenSsidsByName.indexOf(ssidName) >= 0) ? false : this.ssidsState.push(ssidState);
+            const pushName = (this.hidenSsidsName.indexOf(ssidName) >= 0) ? false : this.exposedSsidsName.push(ssidName);
+            const pushState = (this.hidenSsidsName.indexOf(ssidName) >= 0) ? false : this.exposedSsidsState.push(ssidState);
           }
         }
 
-        const exposedSsidsCount = this.ssidsName.length;
+        const exposedSsidsCount = this.exposedSsidsName.length;
         for (let i = 0; i < exposedSsidsCount; i++) {
-          const ssidState = this.ssidsState[i];
+          const ssidState = this.exposedSsidsState[i];
 
           if (this.merakiSsidServices) {
             this.merakiSsidServices[i]
@@ -332,8 +345,8 @@ class merakiDevice {
     this.log.debug('prepareMerakiService');
 
     //get devices data variables
-    const merakiDashboardClientsCount = this.merakiDashboardClientsCount;
-    const allSsidsCount = this.allSsidsCount;
+    const dashboardClientsCount = this.dashboardClientsCount;
+    const ssidsCount = this.ssidsCount;
     const hiddenSsidsCount = this.hiddenSsidsCount;
     const exposedClientByNameCount = this.exposedClientByNameCount;
     const exposedSsidsCount = this.exposedSsidsCount;
@@ -341,24 +354,26 @@ class merakiDevice {
     this.merakiClientPolicyServices = new Array();
     for (let i = 0; i < exposedClientByNameCount; i++) {
       const clientName = this.clientsUser[i];
-      const merakiClientPolicyService = new Service.Switch(clientName, 'merakiClientPolicyService' + i);
+      const clientCustomName = this.getClientByNameOrMac[i].customName;
+      const clientNameToBeExposed = clientCustomName.length > 0 ? clientCustomName : clientName;
+      const merakiClientPolicyService = new Service.Switch(clientNameToBeExposed, 'merakiClientPolicyService' + i);
       merakiClientPolicyService.getCharacteristic(Characteristic.On)
         .onGet(async () => {
           const state = this.clientsPolicyState[i];
           if (!this.disableLogInfo) {
-            this.log('Device: %s, client: %s, get policy, state: %s', accessoryName, clientName, this.clientsPolicyPolicy[i]);
+            this.log('Device: %s, client: %s, get policy, state: %s', accessoryName, clientNameToBeExposed, this.clientsPolicyPolicy[i]);
           }
           return state;
         })
         .onSet(async (state) => {
           const policy = state ? 'Normal' : 'Blocked';
-          const dashboardClientsByIdPolicyUrl = this.host + '/api/v1/networks/' + this.networkId + '/clients/' + this.clientsId[i] + '/policy';
-          const setClientPolicy = this.meraki.put(dashboardClientsByIdPolicyUrl, {
+          const dashboardClientsPolicyUrl = this.host + '/api/v1/networks/' + this.networkId + '/clients/' + this.clientsId[i] + '/policy';
+          const setClientPolicy = this.meraki.put(dashboardClientsPolicyUrl, {
             'devicePolicy': policy
           });
-          this.log.debug('Device: %s %s, debug setClientPolicy: %s', this.host, this.name, setClientPolicy.data);
+          this.log.debug('Device: %s %s, debug setClientPolicy: %s', accessoryName, setClientPolicy.data);
           if (!this.disableLogInfo) {
-            this.log('Device: %s, client: %s, set policy, state: %s', accessoryName, clientName, policy);
+            this.log('Device: %s, client: %s, set policy, state: %s', accessoryName, clientNameToBeExposed, policy);
           }
         });
 
@@ -368,25 +383,25 @@ class merakiDevice {
 
     this.merakiSsidServices = new Array();
     for (let i = 0; i < exposedSsidsCount; i++) {
-      const ssidName = this.ssidsName[i];
+      const ssidName = this.exposedSsidsName[i];
       const merakiSsidService = new Service.Switch(ssidName, 'merakiSsidService' + i);
       merakiSsidService.getCharacteristic(Characteristic.On)
         .onGet(async () => {
-          const state = this.ssidsState[i];
+          const state = this.exposedSsidsState[i];
           if (!this.disableLogInfo) {
-            this.log('Device: %s, SSIDs: %s get state: %s', accessoryName, ssidName, state ? 'ON' : 'OFF');
+            this.log('Device: %s, SSIDs: %s, get state: %s', accessoryName, ssidName, state ? 'ON' : 'OFF');
           }
           return state;
         })
         .onSet(async (state) => {
           state = state ? true : false;
-          let j = this.allSsidsName.indexOf(ssidName);
-          const setSsid = this.meraki.put(this.mrSsidUrl + '/' + [j], {
+          let j = this.ssidsName.indexOf(ssidName);
+          const setSsid = this.meraki.put(this.ssidsUrl + '/' + [j], {
             'enabled': state
           });
-          this.log.debug('Device: %s %s, debug response: %s', this.host, this.name, setSsid.data);
+          this.log.debug('Device: %s %s, debug setSsid: %s', accessoryName, setSsid.data);
           if (!this.disableLogInfo) {
-            this.log('Device: %s, SSID: %s set state: %s', accessoryName, ssidName, state ? 'ON' : 'OFF');
+            this.log('Device: %s, SSID: %s, set state: %s', accessoryName, ssidName, state ? 'ON' : 'OFF');
           }
         });
 
@@ -395,7 +410,7 @@ class merakiDevice {
     }
 
     this.startPrepareAccessory = false;
-    this.log.debug('Device: %s %s, publishExternalAccessories.', this.host, accessoryName);
+    this.log.debug('Device: %s, publishExternalAccessories.', accessoryName);
     this.api.publishExternalAccessories(PLUGIN_NAME, [accessory]);
   }
 }