+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.8.5] - (29.08.2021)
+## Changes
+- added possibility to chose between Name or Mac Adress for clients to be exposed with its poplicy state.
+- added possibility to set custom Name to be exposed for client policy.
+
 ## [0.8.2] - (28.08.2021)
 ## Changes
 - added possibility to expose clients and change its policy.